[ALPS07588156] Coverity: Fix invalid type in argument to printf: 11369436
[Detail]
CID 11369436 (#2 of 2): Invalid type in argument to printf format specifier (PRINTF_ARGS)

[Solution]
Force type conversions to fit format strings

MTK-Commit-Id: ec6a4ceecd207498fc6d677d5bb56237c5a0e148

CR-Id: ALPS07588156
Feature: ARM GPU DDK driver

+++ gpu_mali/mali_valhall/mali-r38p1/drivers/gpu/arm/midgard/platform/mtk_platform_common/mtk_platform_logbuffer.c
@@ -156,11 +156,11 @@ void mtk_logbuffer_print(struct mtk_logbuffer_info *logbuf, const char *fmt, ...
 		 */
 		entry_len = scnprintf(logbuf->tmp_entry,
 		                      MTK_LOG_BUFFER_ENTRY_SIZE, "\n[%5lu.%06lu] %s",
-		                      (unsigned long)ts_nsec, rem_nsec / 1000, buffer);
+		                      (unsigned long)ts_nsec, (unsigned long)rem_nsec / 1000, buffer);
 	} else
 		entry_len = scnprintf(logbuf->tmp_entry,
 		                      MTK_LOG_BUFFER_ENTRY_SIZE, "[%5lu.%06lu] %s",
-		                      (unsigned long)ts_nsec, rem_nsec / 1000, buffer);
+		                      (unsigned long)ts_nsec, (unsigned long)rem_nsec / 1000, buffer);
 
 	if (entry_len == 0 || entry_len >= logbuf->size)
 		goto fail_invalid_entry;
